Fix link showing on contact page.
Addressed the bug which shows an empty icon with a link on the 'contact'
page.  To fix the bug, I added a closing a tag on the 'work' page.

I added an icon for Twitter on my 'contact' page as well.

Future Additions: Professional photo, screenshots/gifs to my work.