Change About Section to Skills
I consolidated the information contained in the about/intro sections and
moved the icons of technical skills I employ in my work to a skills
section.

I believe this adds to the fluidity of my portfolio.

TODO: Add visible text links to demos/github on work section and have a
visible email address on the contact page above the form.